+++ internal/backend/db/db.go
@@ -8,6 +8,7 @@ import (
 	"gorm.io/gorm"
 
 	"go.autokitteh.dev/autokitteh/internal/backend/health/healthreporter"
+	"go.autokitteh.dev/autokitteh/internal/backend/types"
 	"go.autokitteh.dev/autokitteh/sdk/sdkservices"
 	"go.autokitteh.dev/autokitteh/sdk/sdktypes"
 )
@@ -141,10 +142,10 @@ type DB interface {
 
 	// -----------------------------------------------------------------------
 	// TODO(ENG-917): Do not expose scheme outside of DB.
-	SaveSignal(ctx context.Context, signal *Signal) error
-	GetSignal(ctx context.Context, signalID uuid.UUID) (*Signal, error)
+	SaveSignal(ctx context.Context, signal *types.Signal) error
+	GetSignal(ctx context.Context, signalID uuid.UUID) (*types.Signal, error)
 	RemoveSignal(ctx context.Context, signalID uuid.UUID) error
-	ListWaitingSignals(ctx context.Context, dstID sdktypes.EventDestinationID) ([]*Signal, error)
+	ListWaitingSignals(ctx context.Context, dstID sdktypes.EventDestinationID) ([]*types.Signal, error)
 
 	// -----------------------------------------------------------------------
 	SetSecret(ctx context.Context, key string, value string) error
@@ -153,11 +154,4 @@ type DB interface {
 
 	// -----------------------------------------------------------------------
 	GetOwnership(ctx context.Context, entityID sdktypes.UUID) (string, error)
-}
-
-type Signal struct {
-	ID            uuid.UUID
-	WorkflowID    string
-	DestinationID sdktypes.EventDestinationID
-	Filter        string
 }
